lib/rt-dems/requirements.txt
absl-py==2.4.0
ai-edge-litert==2.1.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
astunparse==1.6.3
backports.strenum==1.2.8
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
Django==6.0.3
django-filter==25.2
djangorestframework==3.16.1
fastapi==0.135.1
flatbuffers==25.12.19
gast==0.7.0
google-pasta==0.2.0
grpcio==1.78.0
h11==0.16.0
h5py==3.14.0
idna==3.11
joblib==1.5.3
keras==3.13.2
libclang==18.1.1
lightgbm==4.6.0
markdown-it-py==4.0.0
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
numpy==2.4.3
opt_einsum==3.4.0
optree==0.19.0
packaging==26.0
paho-mqtt==2.1.0
pandas
protobuf==7.34.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
requests==2.32.5
rich==14.3.3
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
six==1.17.0
sqlparse==0.5.5
starlette==0.52.1
termcolor==3.3.0
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wheel==0.46.3
wrapt==2.1.2
